restructured redux folders and destructured authApi using rootApi.js
