--- client/src/redux/api/apiSlice.js
@@ -1,37 +1,0 @@
-import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-
-export const apiSlice = createApi({
-  reducerPath: "api",
-  baseQuery: fetchBaseQuery({
-    baseUrl: "http://localhost:3000/api",
-    prepareHeaders: (headers) => {
-      const token = localStorage.getItem("token");
-      if (token) {
-        headers.set("Authorization", `Bearer ${token}`);
-      }
-      return headers;
-    },
-  }),
-
-  endpoints: (builder) => ({
-    // User Login
-    loginUser: builder.mutation({
-      query: (loginData) => ({
-        url: "/auth/login",
-        method: "POST",
-        body: loginData,
-      }),
-    }),
-
-    // User Registration
-    registerUser: builder.mutation({
-      query: (userData) => ({
-        url: "/auth/register",
-        method: "POST",
-        body: userData,
-      }),
-    }),
-  }),
-});
-
-export const { useLoginUserMutation, useRegisterUserMutation } = apiSlice;
