+++ client/src/redux/auth/authApi.js
@@ -1,0 +1,26 @@
+import { rootApi } from "../api/rootApi";
+
+export const authApi = rootApi.injectEndpoints({
+    
+  endpoints: (builder) => ({
+    // User Login
+    loginUser: builder.mutation({
+      query: (loginData) => ({
+        url: "/auth/login",
+        method: "POST",
+        body: loginData,
+      }),
+    }),
+
+    // User Registration
+    registerUser: builder.mutation({
+      query: (userData) => ({
+        url: "/auth/register",
+        method: "POST",
+        body: userData,
+      }),
+    }),
+  }),
+});
+
+export const { useLoginUserMutation, useRegisterUserMutation } = authApi;
